Use lodestone locking mechanicsm, api-version is 1.16 now
Also polished GUI and make it so that it will support new biomes and materials that are not within 1.16. Fix cooldown bug that locked the locator when it was less than zero seconds. Even not searching 3D, current strategy for fetching biomes works fine for underground biomes such as deep dark, lush caves and dripstone caves.